토큰 유지시간 4시간으로 변경, 프로필 삭제 시 null 처리, shop update null 처리

+++ src/main/java/com/daom/service/FileCloudStorage.java
@@ -73,7 +73,9 @@ public class FileCloudStorage {
     }
 
     public void deleteFile(String savedFileName) {
-        s3Service.delete(savedFileName);
+        if(savedFileName != null){
+            s3Service.delete(savedFileName);
+        }
     }
 
     private String createStoreFileName(String originalFilename) {
